Consolidate GitHub commands
