+++ src/commands/gitHub/connectToGitHub/GitHubRepositoryConnectStep.ts
@@ -6,10 +6,10 @@
 import type { ContainerAppsAPIClient, SourceControl } from "@azure/arm-appcontainers";
 import { AzureWizardExecuteStep, nonNullProp } from "@microsoft/vscode-azext-utils";
 import type { Progress } from "vscode";
-import { ext } from "../../extensionVariables";
-import { createContainerAppsClient } from "../../utils/azureClients";
-import { localize } from "../../utils/localize";
-import { listCredentialsFromRegistry } from "../imageSource/containerRegistry/acr/listCredentialsFromRegistry";
+import { ext } from "../../../extensionVariables";
+import { createContainerAppsClient } from "../../../utils/azureClients";
+import { localize } from "../../../utils/localize";
+import { listCredentialsFromRegistry } from "../../imageSource/containerRegistry/acr/listCredentialsFromRegistry";
 import type { IConnectToGitHubContext } from "./IConnectToGitHubContext";
 
 export class GitHubRepositoryConnectStep extends AzureWizardExecuteStep<IConnectToGitHubContext> {
