+++ src/commands/gitHub/connectToGitHub/IConnectToGitHubContext.ts
@@ -6,8 +6,8 @@
 import type { Registry } from '@azure/arm-containerregistry';
 import type { ExecuteActivityContext, ISubscriptionActionContext } from "@microsoft/vscode-azext-utils";
 import type { AzureSubscription } from "@microsoft/vscode-azureresources-api";
-import type { IGitHubContext } from '../../gitHub/IGitHubContext';
-import type { ContainerAppModel } from "../../tree/ContainerAppItem";
+import type { IGitHubContext } from '../../../gitHub/IGitHubContext';
+import type { ContainerAppModel } from "../../../tree/ContainerAppItem";
 
 export interface IConnectToGitHubContext extends ISubscriptionActionContext, IGitHubContext, ExecuteActivityContext {
     targetContainer: ContainerAppModel;
