+++ src/commands/gitHub/connectToGitHub/ServicePrincipalSecretInputStep.ts
@@ -4,25 +4,25 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AzureWizardPromptStep } from "@microsoft/vscode-azext-utils";
-import { localize } from "../../utils/localize";
-import { validateUtils } from "../../utils/validateUtils";
+import { localize } from "../../../utils/localize";
+import { validateUtils } from "../../../utils/validateUtils";
 import type { IConnectToGitHubContext } from "./IConnectToGitHubContext";
 
-export class ServicePrincipalIdInputStep extends AzureWizardPromptStep<IConnectToGitHubContext> {
+export class ServicePrincipalSecretInputStep extends AzureWizardPromptStep<IConnectToGitHubContext> {
     public async prompt(context: IConnectToGitHubContext): Promise<void> {
-        context.servicePrincipalId = (await context.ui.showInputBox({
-            prompt: localize('servicePrincipalIdPrompt', 'Enter the service principal ID'),
+        context.servicePrincipalSecret = (await context.ui.showInputBox({
+            prompt: localize('servicePrincipalSecretPrompt', 'Enter the service principal secret'),
             validateInput: this.validateInput
         })).trim();
-        context.valuesToMask.push(context.servicePrincipalId);
+        context.valuesToMask.push(context.servicePrincipalSecret);
     }
 
     public shouldPrompt(context: IConnectToGitHubContext): boolean {
-        return !context.servicePrincipalId;
+        return !context.servicePrincipalSecret;
     }
 
-    private validateInput(id: string): string | undefined {
-        id = id ? id.trim() : '';
-        return !validateUtils.isValidLength(id) ? validateUtils.getInvalidLengthMessage() : undefined;
+    private validateInput(secret: string): string | undefined {
+        secret = secret ? secret.trim() : '';
+        return !validateUtils.isValidLength(secret) ? validateUtils.getInvalidLengthMessage() : undefined;
     }
 }
