+++ src/commands/gitHub/disconnectRepo/GitHubRepositoryDisconnectStep.ts
@@ -6,10 +6,10 @@
 import type { ContainerAppsAPIClient, SourceControl } from "@azure/arm-appcontainers";
 import { AzureWizardExecuteStep, nonNullProp } from "@microsoft/vscode-azext-utils";
 import { Progress } from "vscode";
-import { ext } from "../../extensionVariables";
-import { gitHubUrlParse } from "../../gitHub/gitHubUrlParse";
-import { createContainerAppsClient } from "../../utils/azureClients";
-import { localize } from "../../utils/localize";
+import { ext } from "../../../extensionVariables";
+import { gitHubUrlParse } from "../../../gitHub/gitHubUrlParse";
+import { createContainerAppsClient } from "../../../utils/azureClients";
+import { localize } from "../../../utils/localize";
 import { getContainerAppSourceControl } from "../connectToGitHub/getContainerAppSourceControl";
 import { IDisconnectRepoContext } from "./IDisconnectRepoContext";
